back_end/apps/modulo_oficina/ordem_servico/documentos/api/serializers.py
from rest_framework import serializers
from .models import Documento

class DocumentoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Documento
        fields = ['id', 'nome', 'tipo', 'data_inclusao']