back_end/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary  # se usar PostgreSQL, ou omita se usar SQLite